Rename 2018 day1 Go file, fix silly bug
I just called it main.go originally as I hadn't implemented the solution
in any other languages. Now I've got other languages there let's rename
it to be consistent.

And also fix a silly bug which caused the +1 -1 test to fail, but I
ignored it as the others passed _cough_.