2020 day 3 (Go)
A relatively straightforward exercise, although initially I created the
grid as a [][]byte. There was something unusual going on in the grow()
function which would result in the grid doing this (using the example
from the puzzle page):

    Grid before
    ..##.......
    .##........
    ##.........
    #.........#
    .........#.
    ........#..
    .......#..#
    ......#..##
    .....#..##.
    ....#..##.#
    ...#..##.##
    Grid after
    ..##.........##.......
    .##.........##........
    ##.........##.........
    #.........##.........#
    .........#..........#.
    ........#..........#..
    .......#..#.......#..#
    ......#..##......#..##
    .....#..##......#..##.
    ....#..##.#....#..##.#
    ...#..##.##...#..##.##

Somehow it was morphing the grid from a correct arrangement to start
with to this weirdly nearly ordered pattern - maybe just repeating the
first row somehow?

Changing this to []string fixed it without any other changes other than
no longer needing to:

    grid[row] = append(grid[row], grid[row]...)

Something unusual with the way I was appending the []byte?